try to improve

+++ app/src/main/java/com/boolder/boolder/domain/model/Coordinates.kt
@@ -3,4 +3,4 @@ package com.boolder.boolder.domain.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Coordinates(val x: Float, val y: Float)
+data class Coordinates(val x: Double, val y: Double)
